Create rheston_mc.py
